we now have a component
